\name{describe.cloglog.net}
\alias{describe.cloglog.net}
\title{Describe the cloglog.net Zelig Model}
\usage{
  describe.cloglog.net(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the cloglog.net Zelig Model
}

