\name{describe.normal.net}
\alias{describe.normal.net}
\title{Describe the normal.net Zelig Model}
\usage{
  describe.normal.net(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the normal.net Zelig Model
}

