\name{describe.poisson.net}
\alias{describe.poisson.net}
\title{Describe the poisson.net Zelig Model}
\usage{
  describe.poisson.net(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the poisson.net Zelig Model
}

