\name{describe.probit.net}
\alias{describe.probit.net}
\title{Describe the probit.net Zelig Model}
\usage{
  describe.probit.net(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list specifying author, title, etc. information
}
\description{
  Describe the probit.net Zelig Model
}

