\name{describes}
\alias{describe.cloglog.net}
\alias{describe.gamma.net}
\alias{describe.logit.net}
\alias{describe.ls.net}
\alias{describe.normal.net}
\alias{describe.poisson.net}
\alias{describe.probit.net}
\alias{describes}
\title{Describe a `logit' model to Zelig}
\usage{
  \method{describe}{cloglog.net}(...)
    \method{describe}{gamma.net}(...)
    \method{describe}{logit.net}(...)
    \method{describe}{ls.net}(...)
    \method{describe}{normal.net}(...)
    \method{describe}{poisson.net}(...)
    \method{describe}{probit.net}(...)
}
\arguments{
  \item{...}{ignored parameters}
}
\value{
  a list to be processed by `as.description'
}
\description{
  Describe a `logit' model to Zelig
}

