\name{netlm}
\alias{netlm}
\title{Fitting Simple Regression Network Models...}
\usage{netlm(formula, data, ...)}
\description{Fitting Simple Regression Network Models}
\details{Fit social network data.}
\value{Some kind of model}
\arguments{\item{formulA}{formula}
\item{data}{data.frame}
\item{...}{Additional parameters}}
