\name{qi.logit.net}
\alias{qi.logit.net}
\title{Compute Quantities of Interest for the Zelig Model logit.net}
\usage{
  qi.logit.net(obj, x = null, x1 = null, y = null,
    num = 1000, param = null)
}
\arguments{
  \item{obj}{a zelig object}

  \item{x}{a setx object}

  \item{x1}{an optional setx object}

  \item{y}{...}

  \item{num}{an integer specifying the number of
  simulations to compute}

  \item{param}{a parameters object}
}
\value{
  a list of key-value pairs specifying pairing titles of
  quantities of interest with their simulations
}
\description{
  Compute Quantities of Interest for the Zelig Model
  logit.net
}

