\name{summary.netglm}
\alias{summary.netglm}
\title{Network Regression Summaries for Fitted Models...}
\usage{summary.netglm(object, dispersion, correlation=FALSE,
    symbolic.cor=FALSE, ...)}
\description{Network Regression Summaries for Fitted Models}
\details{Return the summary of model fitted using any of the ``netglm'' functions.}
\value{An object of "summary.glm"}
\arguments{\item{object}{A ``netglm'' object}
\item{dispersion}{Specifies the dispersion (???)}
\item{correlation}{Specifies the correlation (???)}
\item{symbolic.cor}{Specifies the symbolic.cor (???)}
\item{...}{Other parameters}}
