\name{vcov.netglm}
\alias{vcov.netglm}
\title{Calculate Variance-Covariance Matrix for a `netglm' Fitted Model Object...}
\usage{\method{vcov}{netglm}(object, ...)}
\description{Calculate Variance-Covariance Matrix for a `netglm' Fitted Model Object}
\details{Returns the variance-covariance matrix for all `netglm' fitted model objects.}
\value{a matrix}
\arguments{\item{obj}{a `netglm' object}
\item{...}{parameters to be passed to `summary.glm'}}
