\name{zelig2gamma.net}
\alias{zelig2gamma.net}
\title{Interface between the Zelig Model gamma...}
\usage{zelig2gamma.net(formula, ..., data)}
\description{Interface between the Zelig Model gamma.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
