\name{zelig2logit.net}
\alias{zelig2logit.net}
\title{Interface between the Zelig Model logit.net and
the Pre-existing Model-fitting Method}
\usage{
  zelig2logit.net(formula, LF = "logit", ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model logit.net and the
  Pre-existing Model-fitting Method
}

