\name{zelig2ls.net}
\alias{zelig2ls.net}
\title{Interface between the Zelig Model ls.net and
the Pre-existing Model-fitting Method}
\usage{
  zelig2ls.net(formula, ..., LF = "identity", data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{LF}{a link function}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model ls.net and the
  Pre-existing Model-fitting Method
}

