\name{zelig2normal.net}
\alias{zelig2normal.net}
\title{Interface between the Zelig Model normal...}
\usage{zelig2normal.net(formula, ..., data)}
\description{Interface between the Zelig Model normal.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
