\name{zelig2probit.net}
\alias{zelig2probit.net}
\title{Interface between the Zelig Model probit...}
\usage{zelig2probit.net(formula, ..., data)}
\description{Interface between the Zelig Model probit.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
