\name{ZeligNetwork-package}
\title{ZeligNetwork...}
\description{ZeligNetwork}
\details{\tabular{ll}{
Package: \tab ZeligNetwork\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligNetwork-package}
\alias{ZeligNetwork}
\docType{package}
\author{UNKNOWN AUTHOR}
\keyword{package}
