\name{summary.netglm}
\alias{summary.netglm}
\title{Network Regression Summaries for Fitted Models}
\usage{
  summary.netglm(object, dispersion = NULL,
    correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
  \item{object}{A ``netglm'' object}

  \item{dispersion}{Specifies the dispersion (???)}

  \item{correlation}{Specifies the correlation (???)}

  \item{symbolic.cor}{Specifies the symbolic.cor (???)}

  \item{...}{Other parameters}
}
\value{
  An object of "summary.glm"
}
\description{
  Return the summary of model fitted using any of the
  ``netglm'' functions.
}

