\name{vcov.netglm}
\alias{vcov.netglm}
\title{Calculate Variance-Covariance Matrix for a `netglm' Fitted Model Object}
\usage{
  \method{vcov}{netglm}(object, ...)
}
\arguments{
  \item{obj}{a `netglm' object}

  \item{...}{parameters to be passed to `summary.glm'}
}
\value{
  a matrix
}
\description{
  Returns the variance-covariance matrix for all `netglm'
  fitted model objects.
}

