\name{zelig2cloglog.net}
\alias{zelig2cloglog.net}
\title{Interface between the Zelig Model cloglog.net and
the Pre-existing Model-fitting Method}
\usage{
  zelig2cloglog.net(formula, ..., LF = NULL, data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model cloglog.net and the
  Pre-existing Model-fitting Method
}

