\name{zelig2cloglog.net}
\alias{zelig2cloglog.net}
\title{Interface between the Zelig Model cloglog...}
\usage{zelig2cloglog.net(formula, ..., data)}
\description{Interface between the Zelig Model cloglog.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
