\name{zelig2gamma.net}
\alias{zelig2gamma.net}
\title{Interface between the Zelig Model gamma.net and
the Pre-existing Model-fitting Method}
\usage{
  zelig2gamma.net(formula, ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model gamma.net and the
  Pre-existing Model-fitting Method
}

