\name{zelig2logit.net}
\alias{zelig2logit.net}
\title{Interface between the Zelig Model logit...}
\usage{zelig2logit.net(formula, ..., data)}
\description{Interface between the Zelig Model logit.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
