\name{zelig2ls.net}
\alias{zelig2ls.net}
\title{Interface between the Zelig Model ls...}
\usage{zelig2ls.net(formula, ..., data)}
\description{Interface between the Zelig Model ls.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
