\name{zelig2poisson.net}
\alias{zelig2poisson.net}
\title{Interface between the Zelig Model poisson...}
\usage{zelig2poisson.net(formula, ..., data)}
\description{Interface between the Zelig Model poisson.net and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
