\name{zelig2probit.net}
\alias{zelig2probit.net}
\title{Interface between the Zelig Model probit.net and
the Pre-existing Model-fitting Method}
\usage{
  zelig2probit.net(formula, LF = "probit", ..., data)
}
\arguments{
  \item{formula}{a formula}

  \item{...}{additonal parameters}

  \item{LF}{a link function. This parameter is not allowed
  for cloglog}

  \item{data}{a data.frame}
}
\value{
  a list specifying '.function'
}
\description{
  Interface between the Zelig Model probit.net and the
  Pre-existing Model-fitting Method
}

