\docType{package}
\name{ZeligNetwork-package}
\alias{ZeligNetwork}
\alias{ZeligNetwork-package}
\title{ZeligNetwork}
\description{
  \tabular{ll}{ Package: \tab ZeligNetwork\cr Version: \tab
  0.1\cr Date: \tab 2011-04-25\cr Depends: Zelig License:
  \tab GPL version 2 or newer\cr }
}
\details{
  Edit this description
}
\author{
  UNKNOWN AUTHOR
}
\keyword{package}

