\docType{package}
\name{ZeligNetwork-package}
\alias{ZeligNetwork}
\alias{ZeligNetwork-package}
\title{ZeligNetwork}
\description{
  ZeligNetwork adds social network regressions to Zelig's
  core functionality. These models are intended for use
  with data-sets that are shaped like three-dimensional
  arrays.
}
\details{
  \tabular{ll}{ Package: \tab ZeligNetwork\cr Version: \tab
  0.5-0\cr Date: \tab 2012-10-22\cr Depends: Zelig (>=
  4.0-11), network, sna\cr License: \tab GPL version 2 or
  newer\cr }
}
\author{
  Matt Owen \email{mowen@iq.harvard.edu}, Skyler J.
  Cranmer, Kosuke Imai, Olivia Lau, and Gary King
}
\keyword{network,}
\keyword{package}
\keyword{regression,}
\keyword{social}
\keyword{Zelig,}

